+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Features
 
 ### Improvements
-* Improve clusterer performance by removing access via dotted fields where possible
+* improve clusterer performance by removing access via dotted fields where possible
 
 ### Bugfix